+++ rayvision_houdini/tests/conftest.py
@@ -1,3 +1,4 @@
+# -*- coding=utf-8 -*-
 """The plugin of the pytest.
 
 The pytest plugin hooks do not need to be imported into any test code, it will
@@ -7,9 +8,8 @@ References:
     https://docs.pytest.org/en/2.7.3/plugins.html
 
 """
+import os, sys
 import pytest
-
-from rayvision_api.task.handle import RayvisionTask
 
 from rayvision_houdini.HfsSql import connectsl
 
@@ -33,23 +33,6 @@ def refer_table():
 
 
 @pytest.fixture()
-def user_info():
-    """user info fixture."""
-    return {
-        "domain": "task.renderbus.com",
-        "platform": "2",
-        "access_id": "df6d1d6s3dc56ds6",
-        "access_key": "fa5sd565as2fd65",
-        "local_os": 'windows',
-        "workspace": "c:/workspace",
-        "render_software": "Houdini",
-        "software_version": "17.5.293",
-        "project_name": "Project1",
-        "plugin_config": {}
-    }
-
-
-@pytest.fixture()
 def cg_file_h(tmpdir):
     """Get render config."""
     return {
@@ -58,20 +41,12 @@ def cg_file_h(tmpdir):
 
 
 @pytest.fixture()
-def task(user_info, cg_file_h, mocker):
-    """Create an RayvisionTask fixture."""
-    mocker_task_id = mocker.patch.object(RayvisionTask, 'get_task_id')
-    mocker_task_id.return_value = '1234567'
-    mocker_user_id = mocker.patch.object(RayvisionTask, 'get_user_id')
-    mocker_user_id.return_value = '10000012'
-    mocker_user_id = mocker.patch.object(RayvisionTask,
-                                         'check_and_add_project_name')
-    mocker_user_id.return_value = '147258'
-    return RayvisionTask(cg_file=cg_file_h['cg_file'], **user_info)
-
-
-@pytest.fixture()
-def houdini(cg_file_h, task):
+def houdini(tmpdir):
     """Create an houdini object fixture."""
-    from rayvision_houdini.cg import Houdini
-    return Houdini(str(cg_file_h['cg_file']), task, 2004, '')
+    from rayvision_houdini.analyze_houdini import AnalyzeHoudini
+    if "win" in sys.platform.lower():
+        os.environ["USERPROFILE"] = str(tmpdir)
+    else:
+        os.environ["HOME"] = str(tmpdir)
+    analyze_houdini = AnalyzeHoudini(str(tmpdir), "17.5.293")
+    return analyze_houdini