+++ CHANGELOG.md
@@ -3,5 +3,9 @@
 * update houdini analyse script ([980c06f](https://github.com/renderbus/rayvision_houdini/commit/980c06f))
 
 
+## 1.1.0 (2020-04-01)
 
+* update requirements.txt.
+* update houdini analyse
+* support for independent analysis